vulkan-cereal: use void* instead of uint64_t for hva
This simplifies the code a bit.  The page-alignment logic is removed
from virtio-gpu-gfxstream-renderer and kept in the Vulkan decoder
-- the original case for that was when the Nvidia Linux driver wasn't
aligning mappings to page bounders.

That should not affect ASG mappings.

BUG=235485545
TEST=compile

+++ host-common/address_space_graphics.cpp
@@ -435,7 +435,7 @@ private:
                     buf = aligned_buf_alloc(ADDRESS_SPACE_GRAPHICS_PAGE_SIZE, create.size);
 
                     struct MemEntry entry = { 0 };
-                    entry.hva = (uint64_t)(uintptr_t)buf;
+                    entry.hva = buf;
                     entry.size = create.size;
                     entry.register_fixed = create.hostmemRegisterFixed;
                     entry.fixed_id = create.hostmemId ? create.hostmemId : 0;
